Init income statement

+++ stock/utils/metric_value_utils.py
@@ -2,4 +2,4 @@ def normalize(metric_value):
     if metric_value == u'N/A':
         return 0
     else:
-        return int(metric_value.replace(",", ""))
+        return float(metric_value.replace(",", ""))
